realizacionde cambios de funcion
