+++ blackjackui.py
@@ -275,8 +275,8 @@ class Ui_Blackjack(object):
         Blackjack.setWindowTitle(_translate("Blackjack", "MainWindow", None))
         self.label_3.setText(_translate("Blackjack", "<html><head/><body><p align=\"center\"><span style=\" font-size:14pt;\">Dinero:</span></p></body></html>", None))
         self.labMoney.setText(_translate("Blackjack", "<html><head/><body><p align=\"center\"><span style=\" font-size:18pt; font-weight:600; color:#4b9600;\">500</span></p></body></html>", None))
-        self.lineBet.setPlaceholderText(_translate("Blackjack", "Enter Bet", None))
-        self.btnBet.setText(_translate("Blackjack", "Enter Bet", None))
+        self.lineBet.setPlaceholderText(_translate("Blackjack", "Pober apuesta", None))
+        self.btnBet.setText(_translate("Blackjack", "Poner apuesta", None))
         self.btnDouble.setText(_translate("Blackjack", "Double", None))
         self.btnHit.setText(_translate("Blackjack", "Hit", None))
-        self.btnStay.setText(_translate("Blackjack", "Stay", None))
+        self.btnStay.setText(_translate("Blackjack", "Quedarse", None))
